Fix w & b diagnostics
This commit fixes two things. First, the coefficients used to
compute surface buoyancy are fixed. They used to be correct
only for one surface; now they are correct for both. Second,
the computation of the part of w that comes from quadratic drag
was not right (wrong sign) and has been fixed.